Modularize batchnorm.
