+++ tfjs-core/src/tensor.ts
@@ -195,12 +195,6 @@ export interface OpHandler {
   concat<T extends Tensor>(tensors: Array<T|TensorLike>, axis: number): T;
   stack<T extends Tensor>(tensors: Array<T|TensorLike>, axis: number): Tensor;
   unstack<T extends Tensor>(value: T, axis: number): Tensor[];
-  batchNorm<R extends Rank>(
-      x: Tensor<R>, mean: Tensor<R>|Tensor1D|TensorLike,
-      variance: Tensor<R>|Tensor1D|TensorLike,
-      offset?: Tensor<R>|Tensor1D|TensorLike,
-      scale?: Tensor<R>|Tensor1D|TensorLike,
-      varianceEpsilon?: number): Tensor<R>;
   all<T extends Tensor>(x: Tensor, axis: number|number[], keepDims: boolean): T;
   any<T extends Tensor>(x: Tensor, axis: number|number[], keepDims: boolean): T;
   logSumExp<T extends Tensor>(
@@ -833,17 +827,6 @@ export class Tensor<R extends Rank = Rank> {
     return this.batchNorm(mean, variance, offset, scale, varianceEpsilon);
   }
 
-  batchNorm(
-      mean: Tensor<R>|Tensor1D|TensorLike,
-      variance: Tensor<R>|Tensor1D|TensorLike,
-      offset?: Tensor<R>|Tensor1D|TensorLike,
-      scale?: Tensor<R>|Tensor1D|TensorLike,
-      varianceEpsilon = .001,
-      ): Tensor<R> {
-    this.throwIfDisposed();
-    return opHandler.batchNorm(
-        this, mean, variance, offset, scale, varianceEpsilon);
-  }
   // Reduction ops.
   all<T extends Tensor>(axis: number|number[] = null, keepDims = false): T {
     this.throwIfDisposed();
